updating augmented row column design, adding graphics for area view
